Add multigpu kmeans fit function
This adds a multigpu kmeans fit function, so that we can use the kmeans clustering
code inside of cuml.

This also adds instantiations for kmeans on int64_t indicies, which unfortunately
also requires int64_t indices for the PW distance functions.

+++ cpp/src/cluster/kmeans_fit_double.cu
@@ -21,25 +21,25 @@ namespace cuvs::cluster::kmeans {
 
 void fit(raft::resources const& handle,
          const cuvs::cluster::kmeans::params& params,
-         raft::device_matrix_view<const float, int> X,
-         std::optional<raft::device_vector_view<const float, int>> sample_weight,
-         raft::device_matrix_view<float, int> centroids,
-         raft::host_scalar_view<float, int> inertia,
+         raft::device_matrix_view<const double, int> X,
+         std::optional<raft::device_vector_view<const double, int>> sample_weight,
+         raft::device_matrix_view<double, int> centroids,
+         raft::host_scalar_view<double, int> inertia,
          raft::host_scalar_view<int, int> n_iter)
 {
-  cuvs::cluster::kmeans::fit<float, int>(
+  cuvs::cluster::kmeans::fit<double, int>(
     handle, params, X, sample_weight, centroids, inertia, n_iter);
 }
 
 void fit(raft::resources const& handle,
          const cuvs::cluster::kmeans::params& params,
-         raft::device_matrix_view<const float, int64_t> X,
-         std::optional<raft::device_vector_view<const float, int64_t>> sample_weight,
-         raft::device_matrix_view<float, int64_t> centroids,
-         raft::host_scalar_view<float, int64_t> inertia,
+         raft::device_matrix_view<const double, int64_t> X,
+         std::optional<raft::device_vector_view<const double, int64_t>> sample_weight,
+         raft::device_matrix_view<double, int64_t> centroids,
+         raft::host_scalar_view<double, int64_t> inertia,
          raft::host_scalar_view<int64_t, int64_t> n_iter)
 {
-  cuvs::cluster::kmeans::fit<float, int64_t>(
+  cuvs::cluster::kmeans::fit<double, int64_t>(
     handle, params, X, sample_weight, centroids, inertia, n_iter);
 }
 }  // namespace cuvs::cluster::kmeans
