+++ cpp/src/cluster/kmeans_predict_double.cu
@@ -21,28 +21,29 @@ namespace cuvs::cluster::kmeans {
 
 void predict(raft::resources const& handle,
              const kmeans::params& params,
-             raft::device_matrix_view<const float, int> X,
-             std::optional<raft::device_vector_view<const float, int>> sample_weight,
-             raft::device_matrix_view<const float, int> centroids,
+             raft::device_matrix_view<const double, int> X,
+             std::optional<raft::device_vector_view<const double, int>> sample_weight,
+             raft::device_matrix_view<const double, int> centroids,
              raft::device_vector_view<int, int> labels,
              bool normalize_weight,
-             raft::host_scalar_view<float> inertia)
+             raft::host_scalar_view<double> inertia)
 
 {
-  cuvs::cluster::kmeans::predict<float, int>(
+  cuvs::cluster::kmeans::predict<double, int>(
     handle, params, X, sample_weight, centroids, labels, normalize_weight, inertia);
 }
+
 void predict(raft::resources const& handle,
              const kmeans::params& params,
-             raft::device_matrix_view<const float, int> X,
-             std::optional<raft::device_vector_view<const float, int>> sample_weight,
-             raft::device_matrix_view<const float, int> centroids,
-             raft::device_vector_view<int64_t, int> labels,
+             raft::device_matrix_view<const double, int64_t> X,
+             std::optional<raft::device_vector_view<const double, int64_t>> sample_weight,
+             raft::device_matrix_view<const double, int64_t> centroids,
+             raft::device_vector_view<int64_t, int64_t> labels,
              bool normalize_weight,
-             raft::host_scalar_view<float> inertia)
+             raft::host_scalar_view<double> inertia)
 
 {
-  cuvs::cluster::kmeans::predict<float, int64_t>(
+  cuvs::cluster::kmeans::predict<double, int64_t>(
     handle, params, X, sample_weight, centroids, labels, normalize_weight, inertia);
 }
 }  // namespace cuvs::cluster::kmeans
